Refactor user-related handlers and routers
