+++ src/routers/users.ts
@@ -1,19 +1,24 @@
 import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { default as express } from 'express';
 import type * as db from '../db';
-import { getUser, getUserAttributes, updateUser } from '../services/users';
-import { getGroupsPerUser } from '../services/groups';
+import {
+  getUserAttributesHandler,
+  getUserGroupsHandler,
+  getUserHandler,
+  getUserOptionsHandler,
+  getUserRegistrationsHandler,
+  updateUserHandler,
+} from '../handlers/users';
 import { isLoggedIn } from '../middleware/isLoggedIn';
-import { getUserOptions } from '../services/options';
-import { getUserRegistrations } from '../services/registrations';
+
 const router = express.Router();
 
 export function usersRouter({ dbPool }: { dbPool: PostgresJsDatabase<typeof db> }) {
-  router.get('/', isLoggedIn(dbPool), getUser(dbPool));
-  router.put('/:userId', isLoggedIn(dbPool), updateUser(dbPool));
-  router.get('/:userId/groups', isLoggedIn(dbPool), getGroupsPerUser(dbPool));
-  router.get('/:userId/attributes', isLoggedIn(dbPool), getUserAttributes(dbPool));
-  router.get('/:userId/options', isLoggedIn(dbPool), getUserOptions(dbPool));
-  router.get('/:userId/registrations', isLoggedIn(dbPool), getUserRegistrations(dbPool));
+  router.get('/', isLoggedIn(dbPool), getUserHandler(dbPool));
+  router.put('/:userId', isLoggedIn(dbPool), updateUserHandler(dbPool));
+  router.get('/:userId/groups', isLoggedIn(dbPool), getUserGroupsHandler(dbPool));
+  router.get('/:userId/attributes', isLoggedIn(dbPool), getUserAttributesHandler(dbPool));
+  router.get('/:userId/options', isLoggedIn(dbPool), getUserOptionsHandler(dbPool));
+  router.get('/:userId/registrations', isLoggedIn(dbPool), getUserRegistrationsHandler(dbPool));
   return router;
 }
