+++ src/services/groups.ts
@@ -27,34 +27,6 @@ export function getGroupsByCategoryId(dbPool: PostgresJsDatabase<typeof db>) {
 }
 
 /**
- * Retrieves groups associated with a specific user.
- * @param dbPool The database connection pool.
- * @returns An asynchronous function that handles the HTTP request and response.
- */
-export function getGroupsPerUser(dbPool: PostgresJsDatabase<typeof db>) {
-  return async function (req: Request, res: Response) {
-    const paramsUserId = req.params.userId;
-    const userId = req.session.userId;
-    if (paramsUserId !== userId) {
-      return res.status(403).json({ errors: ['forbidden'] });
-    }
-    try {
-      const query = await dbPool.query.usersToGroups.findMany({
-        with: {
-          group: true,
-        },
-        where: eq(db.usersToGroups.userId, userId),
-      });
-      const out = query.map((r) => r.group);
-      return res.json({ data: out });
-    } catch (e) {
-      console.log('error getting groups per user ' + JSON.stringify(e));
-      return res.status(500).json({ error: 'internal server error' });
-    }
-  };
-}
-
-/**
  * Retrieves groups associated with a specific user filtered by a group Category ID.
  * @param dbPool The database connection pool.
  * @returns An asynchronous function that handles the HTTP request and response.
