+++ src/services/users.ts
@@ -1,68 +1,10 @@
 import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import * as db from '../db';
-import type { Request, Response } from 'express';
 import { and, eq, ne, or } from 'drizzle-orm';
 import { UserData, insertUserSchema } from '../types/users';
 import { upsertUsersToGroups } from './usersToGroups';
 import { upsertUserAttributes } from './userAttributes';
-
-/**
- * Retrieves user data from the database.
- * @param {PostgresJsDatabase<typeof db>} dbPool - The database connection pool.
- * @returns {Function} - Express middleware function to handle the request.
- */
-export function getUser(dbPool: PostgresJsDatabase<typeof db>) {
-  return async function (req: Request, res: Response) {
-    try {
-      const userId = req.session.userId;
-      const user = await dbPool.query.users.findFirst({
-        where: eq(db.users.id, userId),
-      });
-
-      if (!user) {
-        return res.status(401).json({ errors: ['No user found'] });
-      }
-
-      return res.json({ data: user });
-    } catch (error: unknown) {
-      console.error(`[ERROR] ${JSON.stringify(error)}`);
-      return res.sendStatus(500);
-    }
-  };
-}
-
-/**
- * Retrieves user attributes from the database.
- * @param {PostgresJsDatabase<typeof db>} dbPool - The database connection pool.
- * @returns {Function} - Express middleware function to handle the request.
- */
-export function getUserAttributes(dbPool: PostgresJsDatabase<typeof db>) {
-  return async function (req: Request, res: Response) {
-    try {
-      const userId = req.session.userId;
-      const paramsUserId = req.params.userId;
-
-      if (userId !== paramsUserId) {
-        return res.status(400).json({
-          errors: [
-            {
-              message: 'Not authorized to query this user',
-            },
-          ],
-        });
-      }
-
-      const userAttributes = await dbPool.query.userAttributes.findMany({
-        where: eq(db.userAttributes.userId, userId),
-      });
-
-      return res.json({ data: userAttributes });
-    } catch (error: unknown) {
-      console.error(`[ERROR] ${JSON.stringify(error)}`);
-      return res.sendStatus(500);
-    }
-  };
-}
+import { z } from 'zod';
 
 /**
  * Checks user data for existing entries in the database.
@@ -142,48 +84,26 @@ async function upsertUserData(
  * @param {PostgresJsDatabase<typeof db>} dbPool - The database connection pool.
  * @returns {Function} - Express middleware function to handle the request.
  */
-export function updateUser(dbPool: PostgresJsDatabase<typeof db>) {
-  return async function (req: Request, res: Response) {
-    const queryUserId = req.params.userId;
-    const userId = req.session.userId;
+export async function updateUser(
+  dbPool: PostgresJsDatabase<typeof db>,
+  data: {
+    userId: string;
+    userData: z.infer<typeof insertUserSchema>;
+  },
+) {
+  const { userId, userData } = data;
 
-    if (queryUserId !== userId) {
-      return res.status(400).json({
-        errors: [
-          {
-            message: 'Not authorized to update this user',
-          },
-        ],
-      });
-    }
+  const existingUserErrors = await checkUserData(dbPool, userId, userData);
 
-    const body = insertUserSchema.safeParse(req.body);
+  if (existingUserErrors) {
+    return { errors: existingUserErrors };
+  }
 
-    if (!body.success) {
-      return res.status(400).json({ errors: body.error.errors });
-    }
+  const user = await upsertUserData(dbPool, userId, userData);
 
-    try {
-      const existingUserErrors = await checkUserData(dbPool, userId, body.data);
+  const updatedGroups = await upsertUsersToGroups(dbPool, userId, userData.groupIds);
 
-      if (existingUserErrors) {
-        return res.status(400).json({ errors: existingUserErrors });
-      }
+  const updatedUserAttributes = await upsertUserAttributes(dbPool, userId, userData.userAttributes);
 
-      const user = await upsertUserData(dbPool, userId, body.data);
-
-      const updatedGroups = await upsertUsersToGroups(dbPool, userId, body.data.groupIds);
-
-      const updatedUserAttributes = await upsertUserAttributes(
-        dbPool,
-        userId,
-        body.data.userAttributes,
-      );
-
-      return res.json({ data: { user, updatedGroups, updatedUserAttributes } });
-    } catch (e) {
-      console.error(`[ERROR] ${JSON.stringify(e)}`);
-      return res.sendStatus(500);
-    }
-  };
+  return { data: { user, updatedGroups, updatedUserAttributes } };
 }
